switch to ColumnType, move tests
